requirements.txt
Flask>=2.0
mysql-connector-python>=8.0
Faker>=18.0.0
